temporary commit to checkout another branch, not in working state.

+++ components/gs_effects/include/gs_effect_common_functions.h
@@ -4,6 +4,6 @@
 
 int32_t max(int32_t* x, size_t size);
 
-int sign(int x);
+float sign(int x);
 
 int32_t abs(int32_t x);
